// src/generated/library/Subscription.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Account } from "../live/Account";

export type Subscription = { id: string, account: Account, server: string, label: string, paused: boolean, lastSynced: number | null, lastError: string | null, limited: boolean, };